firefox running bare bones
