+++ source/DasBlog.SmokeTest/IWebRunner.cs
@@ -1,0 +1,7 @@
+﻿namespace DasBlog.SmokeTest
+{
+	public interface IWebRunner
+	{
+		void RunDasBlog();
+	}
+}
