+++ source/DasBlog.SmokeTest/WebRunner.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace DasBlog.SmokeTest
+{
+	internal class WebRunner : IWebRunner
+	{
+		public void RunDasBlog()
+		{
+			throw new NotImplementedException();
+		}
+	}
+}
